+++ src/Core/Helpers/AdminHelper.php
@@ -48,12 +48,6 @@ if (! function_exists('htcms_admin_asset')) {
      * @return boolean
      */
     function htcms_admin_asset($url='') {
-        /*try {
-            return vite($url);
-        } catch (Error $error) {
-            //do nothing
-        }*/
-
         $preFix = Str::contains($url, "?") ? "&" : "?";
         $path = htcms_admin_config('theme_assets');
         $appUrl = env('APP_URL');