+++ src/Database/Seeds/HooksTableSeeder.php
@@ -16,15 +16,15 @@ class HooksTableSeeder extends Seeder
         $table_name = 'hooks';
         $date = date('Y-m-d H:i:s');
         $hooks = array(
-            array('id' => '1','name' => 'Header','alias' => 'HOOK_HEADER','direction' => '','description' => 'This is usually on top of the page','created_at' => $date,'updated_at' => $date,'deleted_at' => NULL),
-            array('id' => '2','name' => 'Hero Section','alias' => 'HOOK_HERO_SECTION','direction' => '','description' => 'Hook for big image or main attraction','created_at' => $date,'updated_at' => $date,'deleted_at' => NULL),
+            array('id' => '1','name' => 'Header','alias' => 'HOOK_HEADER','direction' => NULL,'description' => 'This is usually on top of the page','created_at' => $date,'updated_at' => $date,'deleted_at' => NULL),
+            array('id' => '2','name' => 'Hero Section','alias' => 'HOOK_HERO_SECTION','direction' => NULL,'description' => 'Hook for big image or main attraction','created_at' => $date,'updated_at' => $date,'deleted_at' => NULL),
             array('id' => '3','name' => 'After Hero Section','alias' => 'HOOK_AFTER_HERO_SECTION','direction' => '','description' => 'Hook for just after the hero section','created_at' => $date,'updated_at' => $date,'deleted_at' => NULL),
-            array('id' => '4','name' => 'Left Section 1','alias' => 'HOOK_LEFT_SECTION_ONE','direction' => '','description' => 'Left hook for section one','created_at' => $date,'updated_at' => $date,'deleted_at' => NULL),
-            array('id' => '5','name' => 'Left Section 2','alias' => 'HOOK_LEFT_SECTION_TWO','direction' => '','description' => 'Left hook for section two','created_at' => $date,'updated_at' => $date,'deleted_at' => NULL),
-            array('id' => '6','name' => 'Right','alias' => 'HOOK_RIGHT','direction' => '','description' => 'Right hook','created_at' => $date,'updated_at' => $date,'deleted_at' => NULL),
-            array('id' => '7','name' => 'Middle','alias' => 'HOOK_MIDDLE','direction' => '','description' => 'This hook is usually at the middle','created_at' => $date,'updated_at' => $date,'deleted_at' => NULL),
-            array('id' => '9','name' => 'Footer','alias' => 'HOOK_FOOTER','direction' => '','description' => 'Footer Hook','created_at' => $date,'updated_at' => $date,'deleted_at' => NULL),
-            array('id' => '10','name' => 'One Column','alias' => 'HOOK_ONE_COLUMN','direction' => '','description' => 'This is usually for one column layout','created_at' => NULL,'updated_at' => $date,'deleted_at' => NULL)
+            array('id' => '4','name' => 'Left Section 1','alias' => 'HOOK_LEFT_SECTION_ONE','direction' => NULL,'description' => 'Left hook for section one','created_at' => $date,'updated_at' => $date,'deleted_at' => NULL),
+            array('id' => '5','name' => 'Left Section 2','alias' => 'HOOK_LEFT_SECTION_TWO','direction' => NULL,'description' => 'Left hook for section two','created_at' => $date,'updated_at' => $date,'deleted_at' => NULL),
+            array('id' => '6','name' => 'Right','alias' => 'HOOK_RIGHT','direction' => NULL,'description' => 'Right hook','created_at' => $date,'updated_at' => $date,'deleted_at' => NULL),
+            array('id' => '7','name' => 'Middle','alias' => 'HOOK_MIDDLE','direction' => NULL,'description' => 'This hook is usually at the middle','created_at' => $date,'updated_at' => $date,'deleted_at' => NULL),
+            array('id' => '9','name' => 'Footer','alias' => 'HOOK_FOOTER','direction' => NULL,'description' => 'Footer Hook','created_at' => $date,'updated_at' => $date,'deleted_at' => NULL),
+            array('id' => '10','name' => 'One Column','alias' => 'HOOK_ONE_COLUMN','direction' => NULL,'description' => 'This is usually for one column layout','created_at' => NULL,'updated_at' => $date,'deleted_at' => NULL)
         );