+++ changelog.md
@@ -3,8 +3,10 @@
 #v1.3.8 Changes
 #### Major Update: Please note: This updates need fresh installation.
 - Backend: Vue 2 to Vue 3 - working
+- Backend: Improved Smart Copy Paste - done
+- Backend: `direction` column in `hooks` table has default Null value now. - done 
 - Backend: Move Passport to Sanctum for authentication - done
-- Backend: linked module and live edit bug while adding or editing frontend module.
+- Backend: linked module and live edit bug while adding or editing frontend module. - done
 - Frontend/Backend: Moved to vite - working
 - Frontend: Moved to API based validation. 
 - API: Added v2 api